chore: split indexing mod
